Integral types - Make it so there isn't a weird mix of u32, i16, i8 etc (#1)
* change most i16, u8 etc to i32

* Change some .unwrap() to use match for better error detection

* Change sprite struct to us i32

+++ src/game/mod.rs
@@ -18,8 +18,8 @@ use self::user_interface as ui;
 use std::io::Write;
 use std::io::stdout;
 
-pub const GAME_AREA_X: u8 = 81;
-pub const GAME_AREA_Y: u8 = 45;
+pub const GAME_AREA_X: i32 = 81;
+pub const GAME_AREA_Y: i32 = 45;
 
 pub struct Game {
     renderer: Renderer,
